[dagit] Move data fetching for the asset graph into hooks
